moved flux validate file, changed coco processor to choose first prompt instead of joining,
